moved TileColorer to color_rules.py and added some rules to navigation
