+++ qrogue/game/world/tiles/__init__.py
@@ -1,5 +1,5 @@
 # exporting
-from .tiles import TileCode, TileColorer, Tile
+from .tiles import TileCode, Tile
 from .tiles import Invalid, Debug, Void, Floor, Wall, Obstacle, FogOfWar, Decoration, ControllableTile
 from .walk_trigger_tiles import WalkTriggerTile, Trigger, Teleport, Message, Riddler, ShopKeeper, Collectible, Energy
 from .puzzle_tiles import Enemy, Boss
